+++ src/main/java/com/imb2025/smedico/service/jpa/ConsultorioServiceImpl.java
@@ -2,12 +2,14 @@ package com.imb2025.smedico.service.jpa;
 
 import java.util.List;
 import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.imb2025.smedico.entity.Consultorio;
+import com.imb2025.smedico.exception.ResourceNotFoundException;
 import com.imb2025.smedico.repository.ConsultorioRepository;
 import com.imb2025.smedico.service.IConsultorioService;
+import com.imb2025.smedico.dto.ConsultorioRequestDto;
+
 
 import dto.ConsultorioRequestDTO;
 
@@ -26,14 +28,18 @@ public class ConsultorioServiceImpl implements IConsultorioService {
 	//Buscar por id
 	 @Override
 	 public Consultorio findById(Long id) {
-		 Optional<Consultorio> opt;
-			opt = repository.findById(id);
-			if(opt.isPresent()) {
-				return opt.get();
-			}else {
-				return null;
-		 }
+
+		 return repository.findById(id)
+				    .orElseThrow(() -> new ResourceNotFoundException(
+				        "Consultorio no encontrado con id " + id));			
 	 }
+
+
+    @Override
+    public boolean existsById(Long id) {
+         return repository.existsById(id);
+    }
+
 	 
 	//Listar todos
 	@Override
@@ -44,6 +50,9 @@ public class ConsultorioServiceImpl implements IConsultorioService {
 	// Eliminar por ID
     @Override
     public void deleteById(Long id) {
+    	if(!repository.existsById(id)) {
+    		throw new IllegalArgumentException("El consultorio que desea eliminar no existe");
+    	}
         repository.deleteById(id);
     }
     
@@ -60,9 +69,23 @@ public class ConsultorioServiceImpl implements IConsultorioService {
 	}
     
     @Override
-	public Consultorio fromDto(ConsultorioRequestDTO dto) throws Exception {
-	    
-	    return new Consultorio(dto.getNombre(), dto.getUbicacion(), dto.getPiso());
+	public Consultorio fromDto(ConsultorioRequestDto dto) throws Exception {
+    	if(dto.getNombre() == null || dto.getNombre().isBlank()) {
+    		throw new IllegalArgumentException("El nombre no puede estar vacío");
+    	}
+    	if(dto.getUbicacion() == null || dto.getUbicacion().isBlank()) {
+    		throw new IllegalArgumentException("La ubicación no puede estar vacía");
+    	}
+    	if(dto.getPiso() == 0) {
+    		throw new IllegalArgumentException("El piso no puede ser nulo");
+    	}
+    	Consultorio consultorio = new Consultorio();
+    	
+    	consultorio.setNombre(dto.getNombre());
+    	consultorio.setUbicacion(dto.getUbicacion());
+    	consultorio.setPiso(dto.getPiso());
+    	
+    	return consultorio;
 	}
