+++ src/main/java/com/imb2025/smedico/service/jpa/ResultadoEstudioServiceImpl.java
@@ -1,21 +1,31 @@
 package com.imb2025.smedico.service.jpa;
 
+import com.imb2025.smedico.dto.ResultadoEstudioRequestDto;
+import com.imb2025.smedico.entity.Estudio;
+import com.imb2025.smedico.entity.OrdenEstudio;
+import com.imb2025.smedico.entity.ResultadoEstudio;
+import com.imb2025.smedico.exception.ResourceNotFoundException;
+import com.imb2025.smedico.repository.EstudioRepository;
+import com.imb2025.smedico.repository.OrdenEstudioRepository;
+import com.imb2025.smedico.repository.ResultadoEstudioRepository;
+import com.imb2025.smedico.service.IResultadoEstudioService;
 import java.util.List;
 import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.imb2025.smedico.entity.ResultadoEstudio;
-import com.imb2025.smedico.repository.ResultadoEstudioRepository;
-import com.imb2025.smedico.service.IResultadoEstudioService;
-
 @Service
-public class ResultadoEstudioServiceImpl implements IResultadoEstudioService{
+public class ResultadoEstudioServiceImpl implements IResultadoEstudioService {
 
 	
 	@Autowired
 	private ResultadoEstudioRepository repo;
+	
+	@Autowired
+	private OrdenEstudioRepository ordenEstudioRepository;
+	
+	@Autowired
+	private EstudioRepository estudioRepository;
 	
 	@Override
 	public List<ResultadoEstudio> findAll() {
@@ -24,29 +34,73 @@ public class ResultadoEstudioServiceImpl implements IResultadoEstudioService{
 	}
 
 	@Override
-	public ResultadoEstudio findById(long id) {
-		Optional<ResultadoEstudio> opt;
-		opt = repo.findById(id);
-		
-		if(opt.isPresent()) {
-			return opt.get();
-		}else {
-			return null;
-		}
-		
-		
-	}
+        public ResultadoEstudio findById(Long id) {
+
+		return repo.findById(id)
+			    .orElseThrow(() -> new ResourceNotFoundException(
+			        "Entidad no encontrada con id " + id));
+
+        }
+
+        @Override
+        public boolean existsById(Long id) {
+                return repo.existsById(id);
+        }
+
+        @Override
+        public void deleteById(Long id) {
+                if(!repo.existsById(id)) {
+                        throw new IllegalArgumentException("No se puede eliminar. No existe un resultadoEstudio con ID: "+ id);
+                }
+                repo.deleteById(id);
+
+        }
+	
 
 	@Override
-	public ResultadoEstudio save(ResultadoEstudio resultadoEstudio) {
-		
+	public ResultadoEstudio create(ResultadoEstudio resultadoEstudio) {
+		System.out.println("ResultadoEstudio Guardado: " + resultadoEstudio);	
 		return repo.save(resultadoEstudio);
 	}
 
 	@Override
-	public void deleteById(long id) {
-		
-		repo.deleteById(id);
+        public ResultadoEstudio update(Long id, ResultadoEstudio resultadoEstudio) throws Exception {
+                if (repo.existsById(id)) {
+                        resultadoEstudio.setId(id);
+                        return repo.save(resultadoEstudio);
+                }
+                throw new Exception("No existe el Resultado del Estudio con ID: " + id);
+        }
+
+	@Override
+	public ResultadoEstudio fromDto(ResultadoEstudioRequestDto requestDto) throws Exception {
+	    OrdenEstudio ordenEstudio = ordenEstudioRepository.findById(requestDto.getOrdenEstudioId())
+	        .orElseThrow(() -> new Exception("Orden de Estudio NO encontrado con ID " + requestDto.getOrdenEstudioId()));
+	    
+	    Estudio estudio = estudioRepository.findById(requestDto.getOrdenEstudioId())
+	    	.orElseThrow(() -> new Exception("Estudio NO encontrado con ID " + requestDto.getEstudioId()));
+	    
+	    ResultadoEstudio resultado = new ResultadoEstudio();
+	    resultado.setEstudio(estudio);
+	    resultado.setFechaCarga(null);
+	    resultado.setObservaciones(null);
+	    resultado.setOrdenEstudio(ordenEstudio);	    
+	    return resultado;
+
 	}
 
+
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
+	
 }
